python/requirements.txt
web3==6.10.0
python-dotenv
Flask
click
requests
eth-account
